[misc] removing warning messages. Minimizing interface to wchar_t.
    AMD64: wchar_t is signed
	 ARMv7: wchar_t is  apparently...unsigned
	 Not so portable, even if 4 bytes in both cases.
    Wouldn't 16 bits be enough here ?

+++ toulbar2/src/tb2naryconstr.cpp
@@ -19,7 +19,7 @@ NaryConstraint::NaryConstraint(WCSP *wcsp, EnumeratedVariable** scope_in, int ar
         conflictWeights.push_back(0);
         unsigned int domsize = scope_in[i]->getDomainInitSize();
         tbuf[i] = CHAR_FIRST;
-        if(domsize + CHAR_FIRST > std::numeric_limits<Char>::max()) {
+        if(domsize + CHAR_FIRST > (unsigned int)std::numeric_limits<Char>::max()) {
             cerr << "Nary constraints overflow. Try undefine NARYCHAR in makefile." << endl;
             exit(EXIT_FAILURE);
         }
